+++ key-manager-test/src/test/scala/org/bitcoins/keymanager/bip39/BIP39LockedKeyManagerTest.scala
@@ -1,13 +1,14 @@
-package org.bitcoins.keymanager
+package org.bitcoins.keymanager.bip39
 
 import org.bitcoins.core.crypto.AesPassword
+import org.bitcoins.keymanager.{KeyManagerTestUtil, KeyManagerUnitTest, KeyManagerUnlockError}
 
-class LockedKeyManagerTest extends KeyManagerUnitTest {
+class BIP39LockedKeyManagerTest extends KeyManagerUnitTest {
 
   it must "be able to read a locked mnemonic from disk" in {
     val km = withInitializedKeyManager()
 
-    val unlockedKm = LockedKeyManager.unlock(KeyManagerTestUtil.badPassphrase, km.kmParams) match {
+    val unlockedKm = BIP39LockedKeyManager.unlock(KeyManagerTestUtil.badPassphrase, km.kmParams) match {
       case Right(km) => km
       case Left(err) => fail(s"Failed to unlock key manager ${err}")
     }
@@ -19,7 +20,7 @@ class LockedKeyManagerTest extends KeyManagerUnitTest {
   it must "fail to read bad json in the seed file" in {
     val km = withInitializedKeyManager()
     val badPassword = AesPassword.fromString("other bad password").get
-    LockedKeyManager.unlock(passphrase = badPassword, kmParams = km.kmParams) match {
+    BIP39LockedKeyManager.unlock(passphrase = badPassword, kmParams = km.kmParams) match {
       case Left(KeyManagerUnlockError.BadPassword) => succeed
       case result @ (Left(_) | Right(_)) =>
         fail(s"Expected to fail test with ${KeyManagerUnlockError.BadPassword} got ${result}")
@@ -33,7 +34,7 @@ class LockedKeyManagerTest extends KeyManagerUnitTest {
 
     val badPath = km.kmParams.copy(seedPath = badSeedPath)
     val badPassword = AesPassword.fromString("other bad password").get
-    LockedKeyManager.unlock(badPassword, badPath) match {
+    BIP39LockedKeyManager.unlock(badPassword, badPath) match {
       case Left(KeyManagerUnlockError.MnemonicNotFound) => succeed
       case result @ (Left(_) | Right(_)) =>
         fail(s"Expected to fail test with ${KeyManagerUnlockError.MnemonicNotFound} got ${result}")
