+++ testkit/src/main/scala/org/bitcoins/keymanager/KeyManagerTestUtil.scala
@@ -3,6 +3,7 @@ package org.bitcoins.keymanager
 import java.nio.file.Path
 
 import org.bitcoins.core.crypto.AesPassword
+import org.bitcoins.keymanager.bip39.BIP39KeyManager
 import org.bitcoins.testkit.BitcoinSTestAppConfig
 
 object KeyManagerTestUtil {
@@ -14,5 +15,5 @@ object KeyManagerTestUtil {
       .resolve(WalletStorage.ENCRYPTED_SEED_FILE_NAME)
   }
 
-  val badPassphrase: AesPassword = KeyManager.badPassphrase
+  val badPassphrase: AesPassword = BIP39KeyManager.badPassphrase
 }
