+++ wallet/src/main/scala/org/bitcoins/wallet/api/WalletApi.scala
@@ -16,6 +16,7 @@ import org.bitcoins.core.protocol.transaction.Transaction
 import org.bitcoins.core.protocol.{BitcoinAddress, BlockStamp}
 import org.bitcoins.core.wallet.fee.FeeUnit
 import org.bitcoins.keymanager._
+import org.bitcoins.keymanager.bip39.{BIP39KeyManager, BIP39LockedKeyManager}
 import org.bitcoins.wallet.Wallet
 import org.bitcoins.wallet.config.WalletAppConfig
 import org.bitcoins.wallet.models.{AccountDb, AddressDb, SpendingInfoDb}
@@ -187,7 +188,7 @@ trait LockedWalletApi extends WalletApi {
     val kmParams = walletConfig.kmParams
 
     val unlockedKeyManagerE =
-      LockedKeyManager.unlock(passphrase, kmParams)
+      BIP39LockedKeyManager.unlock(passphrase, kmParams)
     unlockedKeyManagerE match {
       case Right(km) =>
         val w = Wallet(keyManager = km,
@@ -211,7 +212,7 @@ trait LockedWalletApi extends WalletApi {
 
 trait UnlockedWalletApi extends LockedWalletApi {
 
-  def keyManager: KeyManager
+  def keyManager: BIP39KeyManager
 
   /**
     * Locks the wallet. After this operation is called,
